// 4Angular_Movies/angular/src/app/core/models/GetMovie.ts
export interface Movie {

  id: string,
  creationTime: string,
  creatorId: string,
  lastModificationTime: string,
  lastModifierId:string,
  movieAuthor: string,
  title:string,
  genreDescription: string,
  genreid: string,
  authorid: string,
  releaseDate: string,
  authorName: string,
  price: number,
  rating:number,
  runtimeMins: string
  plot: string
  writers: string
  stars: string
  countries: string
  languages: string
  idImdb: string
  awards: string
  linkEmbed: string
  videoDescription: string
  linkPoster1: string
  linkPoster2: string
  image1: string
  image2: string

}